+++ clinicadl/transforms/extraction/__init__.py
@@ -1,5 +1,4 @@
 from .base import Extraction
 from .image import Image
 from .patch import Patch
-from .roi import ROI
 from .slice import Slice
