+++ clinicadl/transforms/extraction/patch.py
@@ -73,7 +73,7 @@ class Patch(Extraction):
     @computed_field
     @property
     def extract_method(self) -> ExtractionMethod:
-        """The method to be used for the extraction process (ROI, Image, Patch, Slice)."""
+        """The method to be used for the extraction process (Image, Patch, Slice)."""
         return ExtractionMethod.PATCH
 
     @field_validator("patch_size", "stride", mode="after")
