+++ clinicadl/transforms/extraction/slice.py
@@ -89,7 +89,7 @@ class Slice(Extraction):
     @computed_field
     @property
     def extract_method(self) -> ExtractionMethod:
-        """The method to be used for the extraction process (ROI, Image, Patch, Slice)."""
+        """The method to be used for the extraction process (Image, Patch, Slice)."""
         return ExtractionMethod.SLICE
 
     @field_validator("borders", mode="after")
