+++ tests/unittests/transforms/extraction/test_image.py
@@ -60,29 +60,6 @@ def test_extract():
     shutil.rmtree(tmp_dir)
 
 
-def test_get_tio_image():
-    image_tensor = torch.randn(1, 3, 4, 5)
-    mask_1 = torch.ones(1, 3, 4, 5)
-    mask_2 = torch.zeros(1, 3, 4, 5)
-    label = torch.ones(1, 3, 4, 5)
-
-    image = Image()
-    tio_image = image.get_tio_image(image_tensor, label, mask_1=mask_1, mask_2=mask_2)
-    assert isinstance(tio_image.image, tio.ScalarImage)
-    assert (tio_image.image.tensor == image_tensor).all()
-    assert isinstance(tio_image.label, tio.LabelMap)
-    assert (tio_image.label.tensor == label).all()
-    assert isinstance(tio_image.mask_1, tio.LabelMap)
-    assert (tio_image.mask_1.tensor == mask_1).all()
-    assert isinstance(tio_image.mask_2, tio.LabelMap)
-    assert (tio_image.mask_2.tensor == mask_2).all()
-
-    tio_image = image.get_tio_image(image_tensor, label=None)
-    assert tio_image.label is None
-    tio_image = image.get_tio_image(image_tensor, label=1)
-    assert tio_image.label == 1
-
-
 def test_extract_tio_sample():
     image = Image()
     image_tensor = torch.randn(1, 3, 4, 5)
