--- tests/unittests/transforms/test_extraction.py
@@ -1,1 +1,0 @@
-# TODO later when we are sure of the preprocessing architecture
